ml-service/requirements.txt
flask==2.2.0
flask-cors==3.0.10
pandas==1.5.0
numpy==1.24.0
vaderSentiment==3.3.2
textblob==0.17.1
scikit-learn==1.2.0
requests==2.28.0
python-dotenv==0.19.0
psycopg2-binary==2.9.3